+++ TVPlayer/android/app/src/main/java/com/magazatvplayer/MainApplication.kt
@@ -38,6 +38,8 @@ class MainApplication : Application(), ReactApplication {
               ReactVideoPackage(),
               OrientationPackage(),
               VectorIconsPackage(),
+              // Custom modules
+              ScreenSharePackage(),
           )
         }